api/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
beautifulsoup4=4.10.0=pypi_0
bertopic=0.9.4=pyhd8ed1ab_0
blas=1.0=mkl
brotlipy=0.7.0=py39hb82d6ee_1004
bzip2=1.0.8=h8ffe710_4
ca-certificates=2021.10.8=h5b45459_0
certifi=2021.10.8=py39hcbf5309_2
cffi=1.15.0=py39h0878f49_0
charset-normalizer=2.0.12=pyhd8ed1ab_0
click=8.1.2=py39hcbf5309_0
colorama=0.4.4=pyh9f0ad1d_0
cryptography=36.0.2=py39h7bc7c5c_1
cycler=0.11.0=pypi_0
cython=0.29.23=pypi_0
dataclasses=0.8=pyhc8e2a94_3
deep-translator=1.8.3=pypi_0
dnspython=2.2.1=pypi_0
filelock=3.6.0=pyhd8ed1ab_0
flask=2.1.1=pyhd8ed1ab_0
fonttools=4.32.0=pypi_0
future=0.18.2=py39hcbf5309_5
gensim=4.1.2=pypi_0
hdbscan=0.8.28=py39h5d4886f_1
huggingface_hub=0.5.0=pyhd8ed1ab_0
idna=3.3=pyhd8ed1ab_0
importlib-metadata=4.11.3=py39hcbf5309_1
importlib_metadata=4.11.3=hd8ed1ab_1
intel-openmp=2022.0.0=h57928b3_3663
itsdangerous=2.1.2=pyhd8ed1ab_0
jinja2=3.1.1=pyhd8ed1ab_0
joblib=1.1.0=pyhd8ed1ab_0
kiwisolver=1.4.2=pypi_0
libblas=3.9.0=12_win64_mkl
libcblas=3.9.0=12_win64_mkl
libffi=3.4.2=h8ffe710_5
liblapack=3.9.0=12_win64_mkl
libuv=1.43.0=h8ffe710_0
libzlib=1.2.11=h8ffe710_1014
llvmlite=0.38.0=py39ha0cd8c8_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=2.1.1=py39hb82d6ee_1
matplotlib=3.5.1=pypi_0
mkl=2021.4.0=h0e2418a_729
mkl-service=2.4.0=py39h6b0492b_0
msys2-conda-epoch=20160418=1
ninja=1.10.2=h2d74725_1
nltk=3.6.7=pyhd8ed1ab_0
numba=0.55.1=py39hb8cd55e_0
numpy=1.21.5=py39h6331f09_0
openssl=1.1.1n=h8ffe710_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.2=py39h2e25243_0
pillow=9.1.0=pypi_0
pip=22.0.4=pyhd8ed1ab_0
plotly=5.7.0=pyhd8ed1ab_0
pycparser=2.21=pyhd8ed1ab_0
pyjwt=2.3.0=pypi_0
pymongo=4.1.1=pypi_0
pynndescent=0.5.6=pyh6c4a22f_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.7=pyhd8ed1ab_0
pysocks=1.7.1=py39hcbf5309_5
python=3.9.12=h9a09f29_1_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytorch=1.10.2=cpu_py39h907fbb5_0
pytz=2022.1=pyhd8ed1ab_0
pyyaml=5.4.1=py39hb82d6ee_1
regex=2022.3.15=py39hb82d6ee_1
requests=2.27.1=pyhd8ed1ab_0
sacremoses=0.0.49=pyhd8ed1ab_0
scikit-learn=1.0.2=py39he931e04_0
scipy=1.8.0=py39hc0c34ad_1
sentence-transformers=2.0.0=pyhd8ed1ab_0
setuptools=62.0.0=py39hcbf5309_0
six=1.16.0=pyh6c4a22f_0
smart-open=5.2.1=pypi_0
soupsieve=2.3.2=pypi_0
sqlite=3.37.1=h8ffe710_0
tbb=2021.5.0=h2d74725_1
tenacity=8.0.1=pyhd8ed1ab_0
textblob=0.17.1=pypi_0
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=h8ffe710_0
tokenizers=0.10.3=py39h8c9392a_1
tqdm=4.64.0=pyhd8ed1ab_0
transformers=4.16.2=pyhd8ed1ab_0
tweet-preprocessor=0.6.0=pypi_0
typing-extensions=4.1.1=hd8ed1ab_0
typing_extensions=4.1.1=pyha770c72_0
tzdata=2022a=h191b570_0
ucrt=10.0.20348.0=h57928b3_0
umap-learn=0.5.2=py39hcbf5309_1
urllib3=1.26.9=pyhd8ed1ab_0
vc=14.2=hb210afc_6
vs2015_runtime=14.29.30037=h902a5da_6
werkzeug=2.1.1=pyhd8ed1ab_0
wheel=0.37.1=pyhd8ed1ab_0
win_inet_pton=1.1.0=py39hcbf5309_4
xz=5.2.5=h62dcd97_1
yaml=0.2.5=h8ffe710_2
zipp=3.8.0=pyhd8ed1ab_0
zlib=1.2.11=h8ffe710_1014
